# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Yasser AI Academy API
  version: 1.0.0
  description: Authenticated education, commerce, portal, tutor, and credential APIs.
servers:
  - url: http://localhost:3001/api
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: yaa_session
...
paths:
  /auth/register:
    post:
      summary: Create a student account and session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName, email, password, acceptedTerms]
              properties:
                fullName: { type: string, minLength: 2 }
                email: { type: string, format: email }
                password: { type: string, minLength: 8, maxLength: 128 }
                acceptedTerms: { const: true }
      responses:
        "201": { description: Account created }
        "409": { description: Email exists }
  /auth/login:
    post:
      summary: Create a session
      responses:
        "200": { description: Authenticated }
        "401": { description: Invalid credentials }
  /auth/logout:
    post:
      security: [{ sessionCookie: [] }]
      responses:
        "204": { description: Session revoked }
  /auth/me:
    get:
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Current user }
        "401": { description: Authentication required }
...
  /auth/reset-password:
    post:
      summary: Consume a one-time reset token and revoke existing sessions
      responses:
        "200": { description: Password replaced }
        "400": { description: Invalid, expired, or consumed token }
  /account/sessions:
    get:
      security: [{ sessionCookie: [] }]
      summary: List active sessions and identify the current session
      responses:
        "200": { description: Session list }
  /account/sessions/{id}:
    delete:
      security: [{ sessionCookie: [] }]
      summary: Revoke one owned session
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Session revoked }
        "404": { description: Session not found }
  /account/export:
    get:
      security: [{ sessionCookie: [] }]
      summary: Export the authenticated user's portable account data
      responses:
        "200": { description: Account export without credential material }
  /account:
    delete:
      security: [{ sessionCookie: [] }]
      summary: Close or anonymize an account after password confirmation
      responses:
        "204": { description: Account closed and sessions revoked }
        "400": { description: Confirmation required }
...
  /certificates/verify/{code}:
    get:
      summary: Public certificate verification
      parameters:
        - in: path
          name: code
          required: true
          schema: { type: string }
      responses:
        "200": { description: Public certificate details }
        "404": { description: Certificate not found }